small code fix
